oops quick reference
These concepts work together to provide a robust framework for designing and organizing code, enhancing flexibility, reusability, and maintainability in object-oriented systems.

Composition is a design principle, simple and easy to understand. It allows to stick with simple, independents components that can be used in composite class which is loosely coupled rather tight coupling like an inheritance.